+++ je_auto_control/utils/table_pattern/__init__.py
@@ -1,0 +1,6 @@
+"""Table headers + cell addressing for native grids (UIA TablePattern / GridItem)."""
+from je_auto_control.utils.table_pattern.table_pattern import (
+    cell_by_header, table_cell, table_headers,
+)
+
+__all__ = ["table_headers", "table_cell", "cell_by_header"]